+++ R/plot_contribution.R
@@ -2,12 +2,8 @@
 #'
 #' This function plots contributions of features into the prediction for a single observation.
 #'
-#' @param shap SHAP values dataframe produced with the \code{treeshap} function, containing only one row.
-#' @param x \code{NULL} or dataframe with 1 observation used to calculate \code{shap}.
-#' Used only for aesthetic reasons - to include observation values for a different variables next to the variable names in labels on the y axis.
-#' By default is \code{NULL} and then labels on the y axis are just variable names.
-#' @param model \code{NULL} or dataframe containing unified representation of explained model created with a (model).unify function.
-#' Used to calculate mean prediction of the model to use as a baseline.
+#' @param treeshap A treeshap object produced with the \code{treeshap} function.
+#' @param obs A numeric indicating which obseravtion should be plotted. Be deafult it's first obseravtion.
 #' If \code{NULL} then baseline will be set as \code{0} and difference between individual prediction and model's mean prediction will be explained.
 #' @param max_vars maximum number of variables that shall be presented. Variables with the highest importance will be presented.
 #' Remaining variables will be summed into one additional contribution. By default \code{5}.
@@ -33,19 +29,23 @@
 #' target <- fifa20$target
 #' param <- list(objective = "reg:squarederror", max_depth = 3)
 #' xgb_model <- xgboost::xgboost(as.matrix(data), params = param, label = target, nrounds = 200)
-#' unified_model <- xgboost.unify(xgb_model)
+#' unified_model <- xgboost.unify(xgb_model, as.matrix(data))
 #' x <- head(data, 1)
 #' shap <- treeshap(unified_model, x)
-#' plot_contribution(shap, x, unified_model, min_max = c(0, 120000000))
+#' plot_contribution(shap, 1,  min_max = c(0, 120000000))
 #' }
-plot_contribution <- function(shap,
-                              x = NULL,
-                              model = NULL,
+plot_contribution <- function(treeshap,
+                              obs = 1,
                               max_vars = 5,
                               min_max = NA,
                               digits = 3,
                               title = "SHAP Break-Down",
                               subtitle = "") {
+
+  shap <- treeshap$treeshap[obs,]
+  model <- treeshap$unified_model
+  x <- treeshap$observations[obs,]
+
   if (max_vars > ncol(shap)) {
     warning("max_vars exceeds number of variables. All variables will be shown.")
     max_vars <- ncol(shap)