+++ R/plot_feature_dependence.R
@@ -2,8 +2,7 @@
 #'
 #' Depending on the value of a variable: how does it contribute into the prediction?
 #'
-#' @param shaps SHAP values dataframe produced with the \code{treeshap} function.
-#' @param x dataset used to calculate \code{shaps}.
+#' @param treeshap A treeshap object produced with the \code{treeshap} function.
 #' @param variable name or index of variable for which feature dependence will be plotted.
 #' @param title the plot's title, by default \code{'Feature Dependence'}.
 #' @param subtitle the plot's subtitle. By default no subtitle.
@@ -27,11 +26,14 @@
 #' unified_model <- xgboost.unify(xgb_model)
 #' x <- head(data, 100)
 #' shaps <- treeshap(unified_model, x)
-#' plot_feature_dependence(shaps, x, variable = "overall")
+#' plot_feature_dependence(shaps, variable = "overall")
 #' }
-plot_feature_dependence <- function(shaps, x, variable,
+plot_feature_dependence <- function(treeshap, variable,
                                     title = "Feature Dependence", subtitle = NULL) {
   # TODO - add interactions as in https://christophm.github.io/interpretable-ml-book/shap.html
+
+  shaps <- treeshap$treeshap
+  x <- treeshap$observations
 
   if (is.character(variable)) {
     if (!(variable %in% colnames(shaps))) {
@@ -53,7 +55,7 @@ plot_feature_dependence <- function(shaps, x, variable,
   }
 
 
-  df <- data.frame(var_value = x[[variable]], shap_value = shaps[[variable]])
+  df <- data.frame(var_value = x[[,variable]], shap_value = shaps[[,variable]])
   p <- ggplot(df, aes(x = var_value, y = shap_value)) +
     geom_point()