+++ tests/testthat/test_recalculate_covers.R
@@ -6,7 +6,7 @@ target <- fifa20$target
 test_that('recalculate covers works correctly for xgboost model', {
   param <- list(objective = "reg:squarederror", max_depth = 5)
   xgb_model <- xgboost::xgboost(as.matrix(data), params = param, label = target, nrounds = 100, verbose = FALSE)
-  unified <- xgboost.unify(xgb_model)
+  unified <- xgboost.unify(xgb_model, as.matrix(data))
   a <- recalculate_covers(unified, data)$Cover
   b <- unified$Cover
   expect_true(all(a == b))